ListView - InputProcess - InputErrors
Removed const from private var in Sensor. TODO-> virtual (PolyMorphism) that can be used to display/calculate each Sensor.
Added ListView on the UI. It will be used for both Active Sensors and to display the available Sensors to choose.
Started to create the InputProcess -> TODO
Defined first Error group (100-199 for Input)